experiments/EXP-003_prompt-search/requirements.txt
# Kräver anthropic SDK (redan i projektets requirements.txt)
# PyYAML för att ladda prompt-definitioner
pyyaml
